[UPDATE] format tanggal, Group Jurnal Harian & Resume Jurnal

+++ app/Http/Livewire/Pelaksanaan/Keuangan/LvJurnalHarian.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Storage;
 use App\Models\{
     Keuangan\JurnalHarian,
+    Keuangan\ResumeJurnal,
 };
 use App\Helpers\StringGenerator;
 
@@ -20,7 +21,7 @@ class LvJurnalHarian extends Component
     ];
     
     public $page_attribute = [
-        'title' => 'Jurnal Harian',
+        'title' => 'Jurnal Keuangan',
     ];
     public $page_permission = [
         'add' => 'jurnal-harian add',
@@ -32,37 +33,58 @@ class LvJurnalHarian extends Component
         'detail' => false,
     ];
     
+    public $tipe_jurnal;
     public $file_image;
     public $input_tanggal;
     public $iteration;
     
     public $items;
-    public $selected_item_group = [];
+    public $selected_item_group = [
+        'resume' => [],
+        'jurnal' => [],
+    ];
     public $selected_group_name;
     public $selected_item;
     public $selected_url;
     
+    public function mount()
+    {
+        $this->tipe_jurnal = 'jurnal';
+    }
+    
     public function render()
     {
+        $resume_items = ResumeJurnal::query()
+        ->select('*')
+        ->selectRaw('DATE_FORMAT(tanggal, "%M %Y") as date, "resume" as type');
+        
         $items = JurnalHarian::query()
         ->select('*')
-        ->selectRaw('DATE_FORMAT(tanggal, "%M %Y") as date')
+        ->selectRaw('DATE_FORMAT(tanggal, "%M %Y") as date, "jurnal" as type')
+        ->unionAll($resume_items)
         ->orderBy('tanggal', 'ASC')
         ->get()
         ->groupBy('date');
-
-
+        
+        
         $this->items = collect($items)->map(function ($values, $index)
         {
+            $data_items = $values->groupBy('type');
             return [
                 'name' => $index,
-                'items' => $values,
+                'items' => [
+                    'resume' => $data_items['resume'] ?? [],
+                    'jurnal' => $data_items['jurnal'] ?? [],
+                ],
             ];
         });
-        
+        // dd($items, $this->items);
         if ($this->selected_group_name) {
             $item = $this->items->where('name', $this->selected_group_name)->first();
-            $this->selected_item_group = $item['items'] ?? [];
+            $this->selected_item_group = $item['items'] ?? [
+                'resume' => [],
+                'jurnal' => [],
+            ];
         }
         
         return view('livewire.pelaksanaan.keuangan.lv-jurnal-harian')
@@ -75,12 +97,22 @@ class LvJurnalHarian extends Component
         $this->validate([
             'file_image' => 'required|image',
             'input_tanggal' => 'required|string',
+            'tipe_jurnal' => 'required|string',
         ]);
-        $date_now = date('Y-m-d H:i:s', strtotime($this->input_tanggal));
+        $model = null;
+        if($this->tipe_jurnal == 'resume') {
+            $model = ResumeJurnal::class;
+        } elseif($this->tipe_jurnal == 'jurnal') {
+            $model = JurnalHarian::class;
+        } else {
+            return $this->dispatchBrowserEvent('notification:show', ['type' => 'error', 'title' => 'Ops!', 'message' => "We tried it, but failed when requesting data to the server, sorry."]);
+        }
+        $date_parse = str_replace('/', '-', $this->input_tanggal);
+        $date_now = date('Y-m-d H:i:s', strtotime($date_parse));
         $image_name = StringGenerator::fileName($this->file_image->extension());
-        $image_path = Storage::disk('sector_disk')->putFileAs(JurnalHarian::BASE_PATH, $this->file_image, $image_name);
+        $image_path = Storage::disk('sector_disk')->putFileAs($model::BASE_PATH, $this->file_image, $image_name);
         
-        $insert = JurnalHarian::create([
+        $insert = $model::create([
             'image_real_name' => $this->file_image->getClientOriginalName(),
             'image_name' => $image_name,
             'tanggal' => $date_now,
@@ -88,7 +120,7 @@ class LvJurnalHarian extends Component
         
         $this->resetInput();
         
-        return $this->dispatchBrowserEvent('notification:success', ['title' => 'Success!', 'message' => 'Successfully adding data.']);
+        return $this->dispatchBrowserEvent('notification:show', ['type' => 'success', 'title' => 'Success!', 'message' => 'Successfully adding data.']);
     }
     
     public function setInputTanggal($value)
@@ -103,11 +135,26 @@ class LvJurnalHarian extends Component
         $this->iteration++;
     }
     
-    public function setItem($id)
+    public function setItem($id, $type)
     {
-        $item = JurnalHarian::findOrFail($id);
-        $this->selected_item = $item;
+        $item = $this->getItemById($id, $type);
+        if(!$item) return $this->dispatchBrowserEvent('notification:show', ['type' => 'warning', 'title' => 'Ops!', 'message' => "Sorry we can't find any data"]);
+        $this->selected_item = $item->toArray();
+        $this->selected_item['type'] = $type;
         $this->selected_url = route('files.image.stream', ['path' => $item->base_path, 'name' => $item->image_name]);
+    }
+    
+    public function getItemById($id, $type)
+    {
+        $model = null;
+        if($type == 'resume') {
+            $model = ResumeJurnal::class;
+        } elseif($type == 'jurnal') {
+            $model = JurnalHarian::class;
+        } else {
+            return false;
+        }
+        return $model::findOrFail($id);
     }
     
     public function setGroupName($name)
@@ -129,15 +176,17 @@ class LvJurnalHarian extends Component
     
     public function downloadImage()
     {
-        $item = JurnalHarian::findOrFail($this->selected_item['id']);
+        $item = $this->getItemById($this->selected_item['id'], $this->selected_item['type']);
+        if(!$item) return $this->dispatchBrowserEvent('notification:show', ['type' => 'warning', 'title' => 'Ops!', 'message' => "Sorry we can't find any data"]);
         $path = $item->full_path;
         
         return Storage::disk('sector_disk')->download($path, $item->image_real_name);
     }
     
-    public function delete($id)
+    public function delete($id, $type)
     {
-        $item = JurnalHarian::findOrFail($id);
+        $item = $this->getItemById($id, $type);
+        if(!$item) return $this->dispatchBrowserEvent('notification:show', ['type' => 'warning', 'title' => 'Ops!', 'message' => "Sorry we can't find any data"]);
         $path = $item->full_path;
         Storage::disk('sector_disk')->delete($path);
         $item->delete();
